db · Frente C · troceo 2/N · extraer snapshot/export/import/reset → services/db/snapshot.ts
Mover-no-reescribir: exportSnapshot, importSnapshot, resetAllData, exportSnapshotJSON,
bulkClearStores + el helper interno exposeAtlasDBHandle (y su llamada module-level que
expone window.atlasDB) salen de db.ts a services/db/snapshot.ts (copia literal). db.ts
re-exporta las 5 funciones públicas → cero cambios para consumidores; el re-export carga
snapshot.ts y ejecuta exposeAtlasDBHandle() igual que antes.

Se exportan desde db.ts DB_NAME y la interfaz AtlasHorizonDB (las necesita snapshot.ts;
export aditivo). JSZip sigue como import dinámico dentro de las funciones.

db.ts 4467→4039 líneas · snapshot.ts 444 · documents.ts 102. tsc 0 · tests de
módulo/snapshot verdes (network-first, initDBBloque3, snapshotDeclaracion, reconciliacion).
